Add buildIsClean arg/flag to control whether entry point generator throws
The entry point generation step of the build throws errors when
it encounters an existing directory outside of src with the same name as
an immediate subdir inside src. This should only occur when the working
directory is in a clean state as otherwise it will error on subsequent
runs as the generation creates those directories outside of src.