Tweak media-editor postbuild script to build both cjs/esm before checking
When running a full bolt build with a single dist type (to reduce build
times), this postbuild script was failing because it needs both cjs/esm
to exist. It also failed to build cjs because of failing at the esm
step.
Now it will gracefully process both dists before checking at the end if
they are valid.