Remove regexp-transformer from build and fix types/jest issue
The transformer is a private package with its own custom build that is
used as a dev tool with no consumers. As such, it does not need to be
part of the build.
The `yarn run sample` command was not working due to a jest typing issue
so that is now fixed as well.