Fix media client TS errors
Media client's tsconfig has an empty `files` field which overrides the
default typings stubs from the root tsconfig. This was done in a
previous PR to use proper fetch-mock typings which are still stubbed
globally.
We therefore need to add the remaining typings modules that still need
to be stubbed to the files field of media-client.