Remove editor-test-helpers build/tsconfig.json
This package is private and therefore does not need to be built for
publication.
Now that enzyme.ts successfully compiles, it produces a
root-level d.ts file since its an entry point. Typescript then uses that
file to resolve 'enzyme' imports rather than the third party module in
node_modules which causes errors.